# docs/contracts/openapi.json
{"openapi":"3.1.0","info":{"title":"BoardGameHub API","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/boardgames/":{"get":{"tags":["BoardGames"],"summary":"List Boardgames","operationId":"list_boardgames_boardgames__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/BoardGameRead"},"type":"array","title":"Response List Boardgames Boardgames  Get"}}}}}},"post":{"tags":["BoardGames"],"summary":"Create Boardgame","operationId":"create_boardgame_boardgames__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BoardGameCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BoardGameRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/boardgames/{boardgame_id}":{"get":{"tags":["BoardGames"],"summary":"Get Boardgame","operationId":"get_boardgame_boardgames__boardgame_id__get","parameters":[{"name":"boardgame_id","in":"path","required":true,"schema":{"type":"integer","title":"Boardgame Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BoardGameRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["BoardGames"],"summary":"Update Boardgame","operationId":"update_boardgame_boardgames__boardgame_id__put","parameters":[{"name":"boardgame_id","in":"path","required":true,"schema":{"type":"integer","title":"Boardgame Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BoardGameUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BoardGameRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["BoardGames"],"summary":"Delete Boardgame","operationId":"delete_boardgame_boardgames__boardgame_id__delete","parameters":[{"name":"boardgame_id","in":"path","required":true,"schema":{"type":"integer","title":"Boardgame Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/stats":{"get":{"tags":["Stats"],"summary":"Get Stats","operationId":"get_stats_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StatsSnapshot"}}}}}}},"/internal/stats/refresh":{"post":{"tags":["Stats"],"summary":"Refresh Stats","operationId":"refresh_stats_internal_stats_refresh_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"Idempotency-Key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Idempotency-Key"}},{"name":"X-Trace-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Trace-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/token":{"post":{"tags":["Auth"],"summary":"Token","operationId":"token_auth_token_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"BoardGameCreate":{"properties":{"name":{"type":"string","title":"Name"},"designer":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Designer"},"year_published":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Year Published"},"min_players":{"type":"integer","title":"Min Players"},"max_players":{"type":"integer","title":"Max Players"},"play_time_min":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Play Time Min"},"complexity":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Complexity"},"rating":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Rating"}},"type":"object","required":["name","min_players","max_players"],"title":"BoardGameCreate"},"BoardGameRead":{"properties":{"name":{"type":"string","title":"Name"},"designer":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Designer"},"year_published":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Year Published"},"min_players":{"type":"integer","title":"Min Players"},"max_players":{"type":"integer","title":"Max Players"},"play_time_min":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Play Time Min"},"complexity":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Complexity"},"rating":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Rating"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["name","min_players","max_players","id"],"title":"BoardGameRead"},"BoardGameUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"designer":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Designer"},"year_published":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Year Published"},"min_players":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Min Players"},"max_players":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Max Players"},"play_time_min":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Play Time Min"},"complexity":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Complexity"},"rating":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Rating"}},"type":"object","title":"BoardGameUpdate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"LoginRequest"},"StatsSnapshot":{"properties":{"total_games":{"type":"integer","title":"Total Games"},"avg_rating":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Avg Rating"},"avg_complexity":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Avg Complexity"},"player_range_counts":{"additionalProperties":{"type":"integer"},"type":"object","title":"Player Range Counts","default":{}},"top_rated":{"items":{"$ref":"#/components/schemas/TopRatedItem"},"type":"array","title":"Top Rated","default":[]},"generated_at":{"type":"string","format":"date-time","title":"Generated At"}},"type":"object","required":["total_games","generated_at"],"title":"StatsSnapshot"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"TokenResponse"},"TopRatedItem":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"rating":{"type":"number","title":"Rating"}},"type":"object","required":["id","name","rating"],"title":"TopRatedItem"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/auth/token"}}}}}}